+++ mfa/configuration.go
@@ -89,7 +89,7 @@ type Configuration struct {
 func NewConfiguration() *Configuration {
 	cfg := &Configuration{
 		DefaultHeader:    make(map[string]string),
-		UserAgent:        "OpenAPI-Generator/0.9.2/go",
+		UserAgent:        "PingOne-GOLANG-SDK/0.10.0/go",
 		Debug:            false,
 		Servers:          ServerConfigurations{
 			{